move components to different folders
