+++ src/components/Slider/Dots.jsx
