+++ src/components/Slider/SliderContent.jsx
