+++ src/components/Slider/index.jsx
