+++ src/components/Slider/slider.css
